Refactor test case
