+++ core/src/test/java/in/testpress/network/TestBaseResourcePager.java
@@ -9,69 +9,18 @@ import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import java.io.IOException;
-import java.util.List;
 
-import in.testpress.core.TestpressCallback;
-import in.testpress.core.TestpressException;
-import in.testpress.core.TestpressRetrofitRequest;
 import in.testpress.core.TestpressSdk;
 import in.testpress.core.TestpressSession;
-import in.testpress.models.AccountActivity;
 import in.testpress.models.InstituteSettings;
-import in.testpress.models.TestpressApiResponse;
+import in.testpress.util.SamplePager;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
-import retrofit2.Response;
-
-import static in.testpress.network.TestpressApiClient.PAGE;
 
 @RunWith(RobolectricTestRunner.class)
 public class TestBaseResourcePager {
     private SamplePager pager;
     private TestpressApiClient apiClient;
-
-    private class SamplePager extends BaseResourcePager<AccountActivity> {
-
-        private TestpressApiClient apiClient;;
-
-        public SamplePager(TestpressApiClient apiClient) {
-            this.apiClient = apiClient;
-        }
-
-        @Override
-        protected Object getId(AccountActivity resource) {
-            return resource.getId();
-        }
-
-        public TestpressRetrofitRequest<AccountActivity> getRetrofitCall(final TestpressApiClient apiClient, final SamplePager pager) {
-            return new TestpressRetrofitRequest<AccountActivity>() {
-                @Override
-                public RetrofitCall<TestpressApiResponse<AccountActivity>> getRetrofitCall(
-                        int page, int size) {
-                    return apiClient.getAccountActivity(pager.queryParams);
-                }
-            };
-        }
-
-        public TestpressCallback<List<AccountActivity>> testpressCallback() {
-            return new TestpressCallback<List<AccountActivity>>() {
-                @Override
-                public void onSuccess(List<AccountActivity> result) {}
-
-                @Override
-                public void onException(TestpressException exception) {}
-            };
-        }
-
-        @Override
-        public Response<TestpressApiResponse<AccountActivity>> getItems(int page, int size) throws IOException {
-            queryParams.put(PAGE, page);
-            queryParams.put("filter", "app");
-            return apiClient.getAccountActivity(queryParams).execute();
-        }
-
-    }
-
     private MockWebServer mockWebServer;
     private static final String USER_TOKEN = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6NDYsInVzZXJfaWQiOjQ2LCJlbWFpbCI6IiIsImV4cCI6MTUxOTAzNjUzM30.FUuyJfYNSAw_VcypZsN8_ZHvZra6gHU3njcXmr-TGVU";
